requirements.txt
# Agent framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0

# Browser automation (login only) + HTTP API client
playwright>=1.40.0
httpx>=0.27.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.0.0

# Telegram bot
python-telegram-bot>=21.0.0

# Scheduling (asyncio scheduler for bot integration)
apscheduler>=3.10.0

# Data & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Financial data (Plaid)
plaid-python>=14.0.0

# Database — PostgreSQL
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Anthropic SDK (direct)
anthropic>=0.25.0
tzdata
